Use PolyKinds in Teletype example, add note to common mistakes
Using PolyKinds obviates the need for explicit kind annotations on
type parameters of kinds other than `*` (e.g. Teletype's `m` monad
paramter). This seems to be the more general and less verbose
solution to the kind error issue noted in the docs; I have added a
comment to this effect.