Proposal: Asynchronous effect
This adds an `Asynchronous` effect that provides an effectful
interface to that provided by Control.Concurrent.Async.

This could be good to include for convenience's sake (since the
`async`, `asyncBound`, and `asyncOn` functions cannot be implemented
solely in terms of `liftIO`), and might also make for a good demo.
On the other hand, it entails a new dependency on `async`, though
`async` is ubiquitous, and does not provide all the functionality
of Control.Concurrent.Async, as the withUnmask family of functions
does not have proper variance.

I am not married to this module living in fused-effects; it's possible
that a dependent package is the best place for this. Whether or not it
belongs in mainline is a philosophical question: do we want to be a
batteries-included framework, or do we want to encourage minimalism?