Add `finally` and `onException` resource effects.
Because a generalized unlifting function is not possible with this
formulation of effects, it falls on us to provide APIs corresponding
to the remaining resource-management functions in `base`. This adds
`finally` and `onException` to `Control.Effect.Resource`, defined in
terms of `bracket` and `bracketOnError` respectively. This brings
`Resource` up to par with the resource-management functions provided
by `Control.Exception`.