Restore but deprecate Eff and legacy handlers.
Because the next release contains significant API churn, we've decided
to make the path easier by restoring but deprecating the legacy idioms
for defining effects. This restores the `handleSum` family of
functions while informing the user that they will be removed in
upcoming releases, as well as defining `Eff m` as a synonym for `m`
and marking it as deprecated.