Add inspection tests to verify fusion properties.
Pulls in the `inspection-testing` package to verify properties of
generated GHC core. These tests verify that carriers are
eliminated (replaced with lambdas), even in the presence of multiple
carriers and right-associative binds.

Hattip (:tophat:) to @isovector, who filed the original issue and
whose tests for polysemy inspired this one heavily.

Fixes #138.