Remove HFunctor's fmap' with GHCs supporting QuantifiedConstraints.
As was mentioned in the `HFunctor` documentation, we can eliminate the
vestigial `fmap'` method with QuantifiedConstraints. This implementation
is provided by Bottu et al.'s _Quantified Class Constraints_.

The advantage of this patch is that it expresses the nature of
`HFunctor` in a more direct manner, rather than with the
dummy-function workaround. The disadvantage is that only GHC 8.6
supports QuantifiedConstraints, so we have to use CPP to define the
polyfill (as it were) to handle it.

This changes the `HFunctor` interface but should maintain
compatibility, as `fmap'` was never implemented by hand outside of `Sum`.