Inline the runState family of functions.
Prior testing (during the implementation of #129) revealed that
applying INLINE pragmas to the `runState` family of functions yielded
slowdowns (though this was not true of `runWriter`, due to `Writer`
being a newtype). The inliner is a very tricky thing. While reading
the `polysemy` source, I found out that its state type inlines
`runState` only at stage 3 and afterwards. It's not clear to me *why*
this produces performance improvements, but it does.

Before:

|        | 100     | 1000   | 10000   |
|--------|---------|--------|---------|
| Writer | 65.58ns | 569ns  | 5500ns |
| State  | 68.39ns | 565ns  | 5628ns |

After:

|        | 100     | 1000   | 10000   |
|--------|---------|--------|---------|
| Writer | 64.93ns | 540ns  | 5225ns  |
| State  | 63.89ns | 540ns  | 5421ns  |

This makes for a roughly 7% performance game under
`-O2`. (Interestingly enough, the performance gains under `-O1` are,
proportionally, much greater.)