Proposal: allow automatic derivation of HFunctor instances.
Generally speaking, the majority of effects do not require scoping,
and as a result we encounter often this definition of `HFunctor`:

```haskell
instance HFunctor MyEffect where
  hmap _ = coerce
```

This definition is valid as long as there are no occurrences of a
higher-order (i.e. scoped) variable in the effect. The `Coercible`
typeclass allows us to express this in the type system:

```haskell
default hmap :: Coercible (h m a) (h n a)
             => (forall x . m x -> n x)
             -> (h m a -> h n a)
hmap _ = coerce
```

By providing this as a valid default implementation of `hmap`, and
enabling `DeriveAnyClass`/`DerivingStrategies` in effect definitions,
we save cruft and boilerplate without any impact to backwards
compatibility or performance.