Add visible type annotation to the README.
I noticed that the README doesn’t have any examples of using `-XTypeApplications` to specialize a `get` or `ask` function, which is a very common paradigm, common enough that we should demonstrate it. I also added a brief explanation of why it’s necessary (though this topic is covered more exhaustively in the `mtl` comparison).