Remove potential for misuse of `runResource`.
As outlined in #180, misuse of the unlifting function that is passed
to `runResource` can lead to surprising behavior, such as silent
dropping of effects. While we could have used the forklift pattern to
perform the bracketing on a separate, IO-bound thread, a simpler
compromise is to make the `withResource` helper function the canonical
way to execute a `Resource` effect.

This is a breaking change, so it will entail that the next release be 0.5.0.0.