Fix strict StateC's Applicative instance.
The presence of `seq` inside a state monad provides speed boosts but
enables unlawful behavior with respect to passing undefined values.
In practice this probably provides a speed improvement, but if you
need to do that you can do it on your own; we're not interested in
providing unlawful monads.